+++ src/sql/nodes/Insert.js
@@ -1,0 +1,20 @@
+'use strict';
+
+const Node = require('../Node');
+
+class Insert extends Node
+{
+	constructor(rows = [])
+	{
+		super();
+
+		this.rows = rows;
+	}
+
+	push(row)
+	{
+		this.rows.push(row);
+	}
+}
+
+module.exports = Insert;