+++ src/sql/nodes/UpdateSet.js
@@ -1,0 +1,21 @@
+'use strict';
+
+const Node = require('../Node');
+
+class UpdateSet extends Node
+{
+	constructor(columnIdent, expression)
+	{
+		super();
+
+		this.columnIdent = columnIdent;
+		this.expression = expression;
+	}
+
+	childNodes()
+	{
+		return [this.columnIdent, this.expression];
+	}
+}
+
+module.exports = UpdateSet;