+++ src/stream/PropertiesPickerTransformer.js
@@ -7,12 +7,13 @@ const DeepCloner = require('../DeepCloner');
 
 class PropertiesPickerTransformer extends JlTransform
 {
-	constructor(paths, makeClone = false)
+	constructor(paths, makeClone = false, updateIfCb = null)
 	{
 		super(JlTransform.ARRAYS_OF_OBJECTS, JlTransform.ARRAYS_OF_OBJECTS);
 
 		this.propertiesPicker = new PropertiesPicker(paths);
 		this.makeClone = !!makeClone;
+		this.updateIfCb = updateIfCb;
 	}
 
 	_transform(chunk, encoding, cb)
@@ -20,16 +21,23 @@ class PropertiesPickerTransformer extends JlTransform
 		var result = [];
 
 		for (let i = 0; i < chunk.length; i++) {
-			if (this.makeClone) {
-				const dest = DeepCloner.clone(chunk[i]);
+			const row = chunk[i];
 
-				this.propertiesPicker.mergeProperties(chunk[i], dest.sources);
+			if (this.updateIfCb && !this.updateIfCb(row)) {
+				result.push(row);
+				continue;
+			}
+
+			if (this.makeClone) {
+				const dest = DeepCloner.clone(row);
+
+				this.propertiesPicker.mergeProperties(row, dest.sources);
 
 				result.push(dest);
 			} else {
 				const dest = new DataRow(null);
 
-				dest.sources = this.propertiesPicker.sliceProperties(chunk[i]);
+				dest.sources = this.propertiesPicker.sliceProperties(row);
 
 				result.push(dest);
 			}